file restructuring + adding in state for type of release note
