+++ src/pages/ReleaseNotes/index.js
@@ -1,44 +1,118 @@
-/* eslint-disable react/no-danger */
-import React from 'react';
+// import { withStyles } from "@mui/styles";
 import {
   withStyles,
 } from '@material-ui/core';
+import { ArrowDropDown } from '@material-ui/icons'; 
+import clsx from 'clsx'; 
+import React, { useState, useEffect } from 'react';
+import jsonLink from '../../bento/releaseNotesData';
 import styles from './styles';
+import ReleaseNotes from './components/releaseNotes';
+import Stats from '../../components/Stats/AllStatsController';
+import { Typography } from '../../components/Wrappers/Wrappers';
+import usePageTitle from '../../components/Analytics/usePageTitle';
 
-const ReleaseNotes = (props) => {
-  const { classes, versionDetails } = props;
+const ReleaseVersions = (props) => {
+  usePageTitle("Release Notes");
+  const { classes } = props;
+  const [jsonData, setJsonData] = useState(null);
+  const [releaseNoteDetails, setReleaseNoteDetails] = useState(null);
+  const [releaseNoteType, setReleaseNoteType] = useState(null);
+  const [expandedSection, setExpandedSection] = useState(null);
+  const [error, setError] = useState(null);
 
-  // Function to convert URLs to anchor tags
-  const convertToAnchorTags = (text) => {
-    const urlRegex = /(https?:\/\/[^\s]+)/g;
-    return text.replace(urlRegex, (url) => `<a href="${url}" target="_blank">${url}</a>`);
+  useEffect(() => {
+    const fetchJsonData = async () => {
+      try {
+        const response = await fetch(jsonLink);
+        const data = await response.json();
+        setJsonData(data);
+        setReleaseNoteDetails(data.VERSIONS[0]); // Set initial version details
+        setReleaseNoteType('data');
+      } catch (error) {
+        setError(error);
+        console.error('Error fetching JSON data:', error);
+      }
+    };
+
+    fetchJsonData();
+  }, []); 
+
+  const handleReleaseNoteClick = (row, type) => {
+    setReleaseNoteDetails(row);
+    setReleaseNoteType(type);
   };
 
   return (
-    <div className={classes.releaseNotesContainer}>
-      <hr className={classes.horizontalLine} />
-      <div className={classes.releaseNotesWrapper}>
-        <h2 className={classes.heading}>{versionDetails.heading}</h2>
-        <p className={classes.releaseDateInfo}>{`Release Date: ${versionDetails.releaseDate}`}</p>
-        <p className={classes.subHeading}>{versionDetails.subHeading}</p>
-        <div className={classes.infoWrapper}>
-          {versionDetails.content.map((item, index) => (
-            <div key={index}>
-              {item.paragraph && <p dangerouslySetInnerHTML={{ __html: convertToAnchorTags(item.paragraph) }} />}
-              {item.list && (
-                <ul>
-                  {item.list.map((listItem, listItemIndex) => (
-                    <li key={listItemIndex} dangerouslySetInnerHTML={{ __html: convertToAnchorTags(listItem) }} />
-                  ))}
-                </ul>
-              )}
+    <>
+      <Stats />
+      <div className={classes.container}>
+        {jsonData && (
+          <>
+            <h1>{jsonData.HEADING}</h1>
+            <div className={classes.wrapper}>
+              <div className={classes.tableWrapper}>
+                <div className={classes.tableExternal}>
+                  <div className={classes.topBorder}/>
+                  <div className={classes.table}>
+                    <div className={classes.tableHead} onClick={() => setExpandedSection(expandedSection === 'data' ? null : 'data')}>
+                      <span className={clsx(classes.releaseHeading, classes.dataHeading)}>  
+                        {"DATA RELEASE NOTES"} 
+                        <ArrowDropDown className={clsx(classes.releaseDropdown, expandedSection === 'data' ? classes.upsideDown : classes.rightsideUp)}/> 
+                      </span>
+                    </div>
+                    <div className={classes.dataRows}>
+                      {jsonData.VERSIONS.map((row) => (
+                        <div key={row.id} onClick={() => handleReleaseNoteClick(row, 'data')} className={expandedSection === 'data' ? classes.visibleRow : classes.hiddenRow}>
+                          <div className={classes.dataVersion} align="left">
+                            <span className={classes.version}>
+                              {"Version: " + row.versionNumber}
+                            </span>
+                          </div> 
+                          <div className={classes.dataDate} align="left">
+                            {"(" + row.releaseDate + ")"}
+                          </div> 
+                        </div>
+                      ))}
+                    </div>
+                    <div className={clsx(classes.tableHead, classes.softwareBorder)} onClick={() => setExpandedSection(expandedSection === 'software' ? null : 'software')}>
+                      <span className={clsx(classes.releaseHeading, classes.softwareHeading)}>  
+                        {"SOFTWARE RELEASE NOTES"} 
+                        <ArrowDropDown className={clsx(classes.releaseDropdown, expandedSection === 'software' ? classes.upsideDown : classes.rightsideUp)}/> 
+                      </span>
+                    </div>
+                    <div className={classes.softwareRows}>
+                      {jsonData.VERSIONS.map((row) => (
+                        <div key={row.id} onClick={() => handleReleaseNoteClick(row, 'software')} className={expandedSection === 'software' ? classes.visibleRow : classes.hiddenRow}>
+                          <div className={classes.dataVersion} align="left">
+                            <span className={classes.version}>
+                              {"Version: " + row.versionNumber}
+                            </span>
+                          </div> 
+                          <div className={classes.dataDate} align="left">
+                            {"(" + row.releaseDate + ")"}
+                          </div> 
+                        </div>
+                      ))}
+                    </div>
+                  </div>
+                </div>
+                <hr className={classes.horizontalLine} />
+              </div>
+              {releaseNoteDetails && <ReleaseNotes releaseNoteDetails={releaseNoteDetails} releaseNoteType={releaseNoteType}/>}
             </div>
-          ))}
-        </div>
+          </>
+        )}
+        {
+          error && (
+            <Typography variant="h5" color="error" size="sm">
+              {error ? `An error has occurred in loading release notes data: ${error}` : 'Recieved wrong data'}
+            </Typography>
+          )
+        }
       </div>
-      <hr className={classes.horizontalLine} />
-    </div>
+    </>
   );
 };
 
-export default withStyles(styles)(ReleaseNotes);
+export default withStyles(styles)(ReleaseVersions);
